UWP review: check to make sure the review scope does not go beyond the foreground window.
Due to late chekcing logic, sometimes another app would be included in the virtual review window (especially this was the case when Edge and the app in question are next to each other in Alt+TAB order). Improved the situation by checking for desktop object as foreground.parent to prevent the reviewer from falling off the edge of the foreground window.